Finished text drawing function
