+++ project/draw_functions.c
@@ -10,12 +10,13 @@
 |Funções para gerar os diversos indicadores da aeronave  |
 +-------------------------------------------------------*/
 
-# include <GL/gl.h>
-# include <SDL2/SDL.h>
-# include <math.h>
+#include <GL/gl.h>
+#include <SDL2/SDL.h>
+#include <math.h>
 
 #include "draw_functions.h"
 #include "global_variables.h"
+#include "font.h"
 
 void draw_artificial_horizon(float pitch, float roll)
 {
@@ -277,306 +278,27 @@ void draw_test(){
     glEnd () ;
 }
 
-void draw_text(){
-	int scale = 10;
-	draw_character(9,3); 
-}
+/* Draw a string of characters using OpenGL and the Press Start font
+ * Accepts the string, its screen coordinates x and y, an angle, and
+ * a scaling value.
+ *
+ * When set to scale 1, each character is 7 pixels tall and wide. */
+void draw_text(char* string, int x, int y, int angle, int scale){
+	//char string[10] = "218417241";
+	int aux = 0;
 
-void draw_square(int x, int y, int size){
-	x = x * size;
-	y = y * size;
-	glBegin(GL_QUADS);
-		glVertex2f(x,y);
-		glVertex2f(x + size,y);
-		glVertex2f(x + size,y + size);
-		glVertex2f(x,y + size);
-	glEnd();
-}
+	glMatrixMode(GL_MODELVIEW);
+	glLoadIdentity();
+	glTranslatef(x,y,0);
+	glRotatef(-angle, 0, 0, 1);
 
-void draw_character(int num, int scale){
-	switch(num){
-		case 0:
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(1,1,scale);
-			draw_square(4,1,scale);
-			draw_square(5,1,scale);
-			draw_square(0,2,scale);
-			draw_square(1,2,scale);
-			draw_square(5,2,scale);
-			draw_square(6,2,scale);
-			draw_square(0,3,scale);
-			draw_square(1,3,scale);
-			draw_square(5,3,scale);
-			draw_square(6,3,scale);
-			draw_square(0,4,scale);
-			draw_square(1,4,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(1,5,scale);
-			draw_square(2,5,scale);
-			draw_square(5,5,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			break;
+	while(string[aux] != '\0'){
+		draw_character(string[aux],scale); 
 
-		case 1:
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(1,1,scale);
-			draw_square(2,1,scale);
-			draw_square(3,1,scale);
-			draw_square(2,2,scale);
-			draw_square(3,2,scale);
-			draw_square(2,3,scale);
-			draw_square(3,3,scale);
-			draw_square(2,4,scale);
-			draw_square(3,4,scale);
-			draw_square(2,5,scale);
-			draw_square(3,5,scale);
-			draw_square(0,6,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			break;
+		if(string[aux+1] != '\0'){
+			glTranslatef(8*scale,0,0);
+		}
 
-		case 2:
-			draw_square(1,0,scale);
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(0,1,scale);
-			draw_square(1,1,scale);
-			draw_square(5,1,scale);
-			draw_square(6,1,scale);
-			draw_square(4,2,scale);
-			draw_square(5,2,scale);
-			draw_square(6,2,scale);
-			draw_square(2,3,scale);
-			draw_square(3,3,scale);
-			draw_square(4,3,scale);
-			draw_square(5,3,scale);
-			draw_square(1,4,scale);
-			draw_square(2,4,scale);
-			draw_square(3,4,scale);
-			draw_square(4,4,scale);
-			draw_square(0,5,scale);
-			draw_square(1,5,scale);
-			draw_square(2,5,scale);
-			draw_square(0,6,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			draw_square(6,6,scale);
-			break;
-
-		case 3:
-			draw_square(1,0,scale);
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(6,0,scale);
-			draw_square(4,1,scale);
-			draw_square(5,1,scale);
-			draw_square(3,2,scale);
-			draw_square(4,2,scale);
-			draw_square(2,3,scale);
-			draw_square(3,3,scale);
-			draw_square(4,3,scale);
-			draw_square(5,3,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(0,5,scale);
-			draw_square(1,5,scale);
-			draw_square(5,5,scale);
-			draw_square(6,5,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			break;
-
-		case 4:
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(2,1,scale);
-			draw_square(3,1,scale);
-			draw_square(4,1,scale);
-			draw_square(5,1,scale);
-			draw_square(1,2,scale);
-			draw_square(2,2,scale);
-			draw_square(4,2,scale);
-			draw_square(5,2,scale);
-			draw_square(0,3,scale);
-			draw_square(1,3,scale);
-			draw_square(4,3,scale);
-			draw_square(5,3,scale);
-			draw_square(0,4,scale);
-			draw_square(1,4,scale);
-			draw_square(2,4,scale);
-			draw_square(3,4,scale);
-			draw_square(4,4,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(4,5,scale);
-			draw_square(5,5,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			break;
-
-		case 5:
-			draw_square(0,0,scale);
-			draw_square(1,0,scale);
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(0,1,scale);
-			draw_square(1,1,scale);
-			draw_square(0,2,scale);
-			draw_square(1,2,scale);
-			draw_square(2,2,scale);
-			draw_square(3,2,scale);
-			draw_square(4,2,scale);
-			draw_square(5,2,scale);
-			draw_square(5,3,scale);
-			draw_square(6,3,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(0,5,scale);
-			draw_square(1,5,scale);
-			draw_square(5,5,scale);
-			draw_square(6,5,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			break;
-
-		case 6:
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(1,1,scale);
-			draw_square(2,1,scale);
-			draw_square(0,2,scale);
-			draw_square(1,2,scale);
-			draw_square(0,3,scale);
-			draw_square(1,3,scale);
-			draw_square(2,3,scale);
-			draw_square(3,3,scale);
-			draw_square(4,3,scale);
-			draw_square(5,3,scale);
-			draw_square(0,4,scale);
-			draw_square(1,4,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(0,5,scale);
-			draw_square(1,5,scale);
-			draw_square(5,5,scale);
-			draw_square(6,5,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			break;
-
-		case 7:
-			draw_square(0,0,scale);
-			draw_square(1,0,scale);
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(6,0,scale);
-			draw_square(0,1,scale);
-			draw_square(1,1,scale);
-			draw_square(5,1,scale);
-			draw_square(6,1,scale);
-			draw_square(4,2,scale);
-			draw_square(5,2,scale);
-			draw_square(3,3,scale);
-			draw_square(4,3,scale);
-			draw_square(2,4,scale);
-			draw_square(3,4,scale);
-			draw_square(2,5,scale);
-			draw_square(3,5,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			break;
-
-		case 8:
-			draw_square(1,0,scale);
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(0,1,scale);
-			draw_square(1,1,scale);
-			draw_square(5,1,scale);
-			draw_square(0,2,scale);
-			draw_square(1,2,scale);
-			draw_square(2,2,scale);
-			draw_square(5,2,scale);
-			draw_square(1,3,scale);
-			draw_square(2,3,scale);
-			draw_square(3,3,scale);
-			draw_square(4,3,scale);
-			draw_square(0,4,scale);
-			draw_square(3,4,scale);
-			draw_square(4,4,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(0,5,scale);
-			draw_square(5,5,scale);
-			draw_square(6,5,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			draw_square(5,6,scale);
-			break;
-
-		case 9:
-			draw_square(1,0,scale);
-			draw_square(2,0,scale);
-			draw_square(3,0,scale);
-			draw_square(4,0,scale);
-			draw_square(5,0,scale);
-			draw_square(0,1,scale);
-			draw_square(1,1,scale);
-			draw_square(5,1,scale);
-			draw_square(6,1,scale);
-			draw_square(0,2,scale);
-			draw_square(1,2,scale);
-			draw_square(5,2,scale);
-			draw_square(6,2,scale);
-			draw_square(1,3,scale);
-			draw_square(2,3,scale);
-			draw_square(3,3,scale);
-			draw_square(4,3,scale);
-			draw_square(5,3,scale);
-			draw_square(6,3,scale);
-			draw_square(5,4,scale);
-			draw_square(6,4,scale);
-			draw_square(4,5,scale);
-			draw_square(5,5,scale);
-			draw_square(1,6,scale);
-			draw_square(2,6,scale);
-			draw_square(3,6,scale);
-			draw_square(4,6,scale);
-			break;
+		aux++;
 	}
 }
